Separated strats pipeline to a base class
